Code refactoring, moved the registration functions from the blenderRCBPPanel.py to the __init__.py file. This allows blender to load the add-on when it is started.
