+++ script/blenderRCBPanel.py
@@ -4,25 +4,11 @@
 # This software may be modified and distributed under the terms of the
 # BSD-3-Clause license. See the accompanying LICENSE file for details.
 
-bl_info = {
-    "name": "Robotics Utils",
-    "description": "",
-    "author": "Nicogene",
-    "version": (0, 0, 1),
-    "blender": (2, 93, 3),
-    "location": "3D View > Tools",
-    "warning": "",  # used for warning icon and text in addons panel
-    "wiki_url": "",
-    "tracker_url": "",
-    "category": "Development"
-}
-
-
 import bpy
 import os
-import sys
+# import sys
 import yarp
-import numpy as np
+# import numpy as np
 import math
 import json
 from bpy_extras.io_utils import ImportHelper
@@ -401,55 +387,3 @@ class OT_OpenConfigurationFile(Operator, ImportHelper):
         self.parse_conf(self.filepath, context)
         return {'FINISHED'}
     
-# ------------------------------------------------------------------------
-#    Registration
-# ------------------------------------------------------------------------
-
-classes = (
-    MyProperties,
-    WM_OT_Disconnect,
-    WM_OT_Connect,
-    WM_OT_Configure,
-    OBJECT_PT_robot_controller,
-    OT_OpenConfigurationFile,
-    ListItem,
-    MY_UL_List
-)
-
-def register():
-    for cls in classes:
-        bpy.utils.register_class(cls)
-
-    bpy.types.Scene.my_tool = PointerProperty(type=MyProperties)
-    bpy.types.Scene.my_list = CollectionProperty(type=ListItem)
-    bpy.types.Scene.list_index = IntProperty(name="Index for my_list",
-                                             default=0)
-
-    # initialize the dict
-    bpy.types.Scene.rcb_wrapper = {}
-
-    # init the callback
-    bpy.app.handlers.frame_change_post.append(move)
-
-def unregister():
-    for cls in reversed(classes):
-        try:
-            bpy.utils.unregister_class(cls)
-        except:
-            print("An exception was raised when unregistering the class")
-    try:
-        del bpy.types.Scene.my_tool
-        del bpy.types.Scene.my_list
-        del bpy.types.Scene.list_index
-    except:
-        print("Exception raised when deleting the scene.")
-
-    try:
-        # remove the callback
-        bpy.app.handlers.frame_change_post.clear()
-    except:
-        print("Exception raised when removing the callback")
-
-
-if __name__ == "__main__":
-    register()
